[backend] Revert cache package rename
